// frontend/src/app/api/model/bookSummaryDto.ts
/**
 * Book Shelf API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookSummaryDto {
  /**
   * Unique book identifier.
   */
    id?: string;
  /**
   * Book title.
   */
    title?: string;
  /**
   * Primary author\'s display name.
   */
    authorName?: string;
  /**
   * URL of the cover image; empty when none is set.
   */
    coverImageUrl?: string;
}